added clientBidCard and voucherBidCard.js components.  This is what the feed will create once a fetch is done
